fix: pass signed file position when moving file pointer

+++ src/em_rw_interfaces.h
@@ -13,7 +13,8 @@ struct IReadSeekableWrapper : public emscripten::wrapper<parakeet_crypto::IReadS
     }
     void Seek(size_t position, parakeet_crypto::SeekDirection seek_dir)
     {
-        return call<void>("Seek", position, seek_dir);
+        // FIXME: Do we need to support files larger than 4GB?
+        return call<void>("Seek", (int32_t)position, (int)seek_dir);
     }
     size_t GetSize()
     {
